+++ crates/biome_js_parser/src/lexer/mod.rs
@@ -1034,6 +1034,7 @@ impl<'src> JsLexer<'src> {
             b"of" => OF_KW,
             b"out" => OUT_KW,
             b"using" => USING_KW,
+            b"defer" => DEFER_KW,
             _ => T![ident],
         }
     }